requirement.txt
mysql-connector-python
google-generativeai
python-dotenv
SpeechRecognition
IPython
gtts==2.5.0
pyaudio
pyttsx3
pywhatkit
datetime
wikipedia
pyjokes
streamlit
pydub
pygame
sqlalchemy
google_images_search
elevenlabs